fix(range-selector): disable context menu
close https://github.com/dream-num/univer-pro/issues/2731
